+++ src/main/java/com/gmail/nossr50/skills/unarmed/Unarmed.java
@@ -10,9 +10,9 @@ import com.gmail.nossr50.config.AdvancedConfig;
 import com.gmail.nossr50.config.Config;
 
 public class Unarmed {
-    public static int ironArmMinBonusDamage = AdvancedConfig.getInstance().getIronArmMinBonus();
-    public static int ironArmMaxBonusDamage = AdvancedConfig.getInstance().getIronArmMaxBonus();
-    public static int ironArmIncreaseLevel  = AdvancedConfig.getInstance().getIronArmIncreaseLevel();
+    public static double ironArmMinBonusDamage = AdvancedConfig.getInstance().getIronArmMinBonus();
+    public static double ironArmMaxBonusDamage = AdvancedConfig.getInstance().getIronArmMaxBonus();
+    public static int    ironArmIncreaseLevel  = AdvancedConfig.getInstance().getIronArmIncreaseLevel();
 
     public static int    disarmMaxBonusLevel = AdvancedConfig.getInstance().getDisarmMaxBonusLevel();
     public static double disarmMaxChance     = AdvancedConfig.getInstance().getDisarmChanceMax();
